requirements.txt
# Core FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# AI/ML Libraries
sentence-transformers>=2.2.2
pinecone-client>=3.0.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0

# Data handling
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# Web and utilities
python-multipart>=0.0.6
jinja2>=3.1.0
aiofiles>=23.0.0
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
httpx>=0.25.0 